changes some ints to size_ts

+++ src/ncextra.cpp
@@ -49,7 +49,7 @@ std::vector<float> wgetvecvars(WINDOW * scr,std::vector<std::string> ar) //Strin
 	size_t length = ar.size(); 			//get size of input vector
 	std::vector<float> x(length);		//the float vector that we return
 	char word[20];						//char array to get inputs from
-	for (int i = 0; i< length; i++)
+	for (size_t i = 0; i< length; i++)
 	{
 		clear();
 		wprintnccenter(scr,ar[i].c_str()); 	//print a message to the center of the screen
@@ -73,7 +73,7 @@ std::vector<float> wgetfancyvars(WINDOW *scr, std::vector<std::string> strin)
 	wprintnccenter(scr,1,0,strin[0].c_str());
 	std::vector<float> x;
 	x.resize(length-1);
-	for (int i = 1; i < length; i++)
+	for (size_t i = 1; i < length; i++)
 	{
 		wprintnccenter(scr,-i,0,strin[i].c_str());
 		wgetstr(scr,in);
